+++ testing/test_package/lib/example.dart
@@ -2,11 +2,13 @@
 library ex;
 
 import 'dart:async';
+import 'dart:math';
 
 import 'src/mylib.dart' show Helper;
 import 'package:test_package_imported/main.dart';
 
 export 'dart:core' show deprecated, Deprecated;
+import 'package:meta/meta.dart' show protected, factory;
 
 export 'fake.dart' show Cool;
 export 'src/mylib.dart' show Helper;
@@ -27,6 +29,14 @@ const incorrectDocReference = 'same name as const from fake';
 
 /// This should [not work].
 const incorrectDocReferenceFromEx = 'doh';
+
+/// A custom annotation.
+class aThingToDo {
+  final String who;
+  final String what;
+
+  const aThingToDo(this.who, this.what);
+}
 
 const ConstantCat MY_CAT = const ConstantCat('tabby');
 const List<String> PRETTY_COLORS = const <String>[
@@ -233,6 +243,12 @@ class Dog implements Cat, E {
   @deprecated
   int deprecatedField;
 
+  final int aFinalField;
+  static const String aStaticConstField = "A Constant Dog";
+
+  @protected
+  final int aProtectedFinalField;
+
   Dog() {
     testMethod([]);
   }
@@ -249,6 +265,8 @@ class Dog implements Cat, E {
   @override
   bool get isImplemented => true;
 
+  int get aGetterReturningRandomThings => (new Random()).nextInt(50);
+
   @override
   operator ==(other) => other is Dog && name == other.name;
 
@@ -257,6 +275,20 @@ class Dog implements Cat, E {
   @deprecated
   List<Apple> getClassA() {
     return [new Apple()];
+  }
+
+  @Deprecated('before v27.3')
+  List<Dog> getAnotherClassD() {
+    return [new Dog()];
+  }
+
+  /// A tasty static + final property.
+  static final int somethingTasty;
+
+  static int __staticbacker = 0;
+  static int get staticGetterSetter => __staticbacker;
+  static int set staticGetterSetter(x) {
+    __staticbacker = x;
   }
 
   /// Macro method
@@ -311,9 +343,21 @@ class Klass {
   /// A method
   method() {}
 
+  /// A protected method
+  @protected
+  imProtected() {}
+
+  /// Not really a factory, but...
+  @factory
+  imAFactoryNoReally() {}
+
   /// A shadowed method
   @override
   toString() {}
+
+  /// A method with a custom annotation
+  @aThingToDo('from', 'thing')
+  anotherMethod() {}
 }
 
 class MyError extends Error {}